+++ test/wrap-file-ownership.test.js
@@ -24,7 +24,7 @@ const commitStep = require('../lib/wrap-steps/commit');
 const coverage = require('../lib/wrap-steps/changelog-coverage');
 const launchBaseline = require('../lib/launch-baseline');
 const wrapScope = require('../lib/wrap-scope');
-const { execFileArgs } = require('../lib/wrap-steps/_exec-shell');
+const { execFileArgs } = require('../lib/exec');
 
 const ENV = { ...process.env, GIT_AUTHOR_NAME: 't', GIT_AUTHOR_EMAIL: 't@t', GIT_COMMITTER_NAME: 't', GIT_COMMITTER_EMAIL: 't@t' };
 const dirs = [];